(FACT-2824) Extract virtual methods in a separate class.
